# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tu Tiên Lộ Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get my inventory
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/{itemId}/use:
    post:
      operationId: useItem
      tags: [inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      summary: Use an item
      responses:
        "200":
          description: Item used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseItemResult"
  /inventory/{itemId}/equip:
    post:
      operationId: equipItem
      tags: [inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      summary: Equip an item
      responses:
        "200":
          description: Equipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /inventory/{itemId}/unequip:
    post:
      operationId: unequipItem
      tags: [inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      summary: Unequip an item
      responses:
        "200":
          description: Unequipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /inventory/{itemId}/sell:
    post:
      operationId: sellItem
      tags: [inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      summary: Sell an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellItemBody"
      responses:
        "200":
          description: Sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /market:
    get:
      operationId: getMarketListings
      tags: [market]
      summary: Get market listings
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketListing"
  /market/list:
    post:
      operationId: createMarketListing
      tags: [market]
      summary: List item on market
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
  /market/{listingId}/buy:
    post:
      operationId: buyMarketItem
      tags: [market]
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: string
      summary: Buy a market listing
      responses:
        "200":
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    InventoryItem:
      type: object
      properties:
        id:
          type: string
        templateId:
          type: string
        name:
          type: string
        type:
          type: string
        quality:
          type: string
          enum: [pham, linh, huyen, tien, than]
        qty:
          type: integer
        enhanceLv:
          type: integer
        equipped:
          type: boolean
        slot:
          type: string
          nullable: true
        iconUrl:
          type: string
          nullable: true
        description:
          type: string
        baseStats:
          type: object
          additionalProperties: true
      required:
        - id
        - templateId
        - name
        - type
        - quality
        - qty
        - enhanceLv
        - equipped
        - description
        - baseStats
    UseItemResult:
      type: object
      properties:
        message:
          type: string
        effects:
          type: array
          items:
            type: string
      required:
        - message
        - effects
    SellItemBody:
      type: object
      properties:
        qty:
          type: integer
          minimum: 1
      required:
        - qty
...
    MarketListing:
      type: object
      properties:
        id:
          type: string
        sellerName:
          type: string
        itemName:
          type: string
        itemType:
          type: string
        quality:
          type: string
        price:
          type: integer
        qty:
          type: integer
        enhanceLv:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerName
        - itemName
        - itemType
        - quality
        - price
        - qty
        - enhanceLv
        - description
        - createdAt
    CreateListingBody:
      type: object
      properties:
        inventoryItemId:
          type: string
        price:
          type: integer
          minimum: 1
        qty:
          type: integer
          minimum: 1
      required:
        - inventoryItemId
        - price
        - qty
...